[feat] #42 시트 동적높이 설정 modifier 구성
