+++ Projects/DSKit/Sources/Components/PokitDeleteBottomSheet.swift
@@ -65,36 +65,6 @@ public struct PokitDeleteBottomSheet: View {
         .presentationDetents([.height(self.height)])
     }
 }
-
-private struct ReadHeightModifier: ViewModifier {
-    private var sizeView: some View {
-        GeometryReader { geometry in
-            Color.clear.preference(
-                key: HeightPreferenceKey.self,
-                value: geometry.size.height
-            )
-        }
-    }
-    
-    func body(content: Content) -> some View {
-        content.background(sizeView)
-    }
-}
-
-private extension View {
-    func readHeight() -> some View {
-        self.modifier(ReadHeightModifier())
-    }
-}
-
-private struct HeightPreferenceKey: PreferenceKey {
-    static var defaultValue: CGFloat?
-    
-    static func reduce(value: inout CGFloat?, nextValue: () -> CGFloat?) {
-        guard let nextValue = nextValue() else { return }
-        value = nextValue
-    }
-}
 //MARK: - Delegate
 public extension PokitDeleteBottomSheet {
     enum SheetType {
